Добавлены UIButton и UILabel extension, добавлена локализация строк в корзине
