+++ FakeNFT/Resources/en.lproj/Localizable.strings
@@ -13,7 +13,20 @@
 "cart.price_label_name" = "Price";
 "cart.payment_button_title" = "Pay";
 "cart.getOrderError" = "Unfortunately, the data could not be loaded. Please try again later.";
+"cart.cartScreen.emptyCartMessage"  = "Cart is empty";
 
+"cart.payResultScreen.succsesPaymentText" = "Success! The payment has been completed,\ncongratulations on your purchase!";
+"cart.payResultScreen.buttonText" = "Return to catalog";
+"cart.payResultScreen.errorAlertTitle" = "Oops! Something went wrong :(";
+"cart.payResultScreen.errorAlertMessage" = "Try again!";
+
+"cart.PayScreen.screenHeader" = "Select a Payment Method";
+"cart.PayScreen.termOfUseFirstString" = "By making a purchase, you agree to the terms and conditions";
+"cart.PayScreen.termOfUseSecondString" ="User Agreement";
+
+"cart.DeleteConfirmScreen.confiremText" =  "Are you sure you want to remove the item from the cart?";
+"cart.DeleteConfirmScreen.returnButtonTitle" = "Return";
+"cart.DeleteConfirmScreen.deleteButtonTitle" = "Delete";
 
 "alert.header" = "Message for you!";
 "alert.button" = "Ok";
