+++ FakeNFT/Scenes/Cart/DataProvider/PayDataProvider.swift
@@ -12,35 +12,26 @@ protocol PayDataProviderProtocol {
     func payOrder(currencyId: String, _ completion: @escaping (Result<PaymentDto, Error>) -> Void)
 }
 
-
-struct CurrenciesRequest: NetworkRequest {
-    var endpoint: URL? = URL(string: "https://651ff0cc906e276284c3c1bc.mockapi.io/api/v1/currencies")
-}
-
-struct PayRequest: NetworkRequest {
-    // var httpMethod: HttpMethod { .put }
-    // var dto: Encodable?
-    var endpoint: URL? = URL(string: "https://651ff0cc906e276284c3c1bc.mockapi.io/api/v1/orders/1")
-    
-    init(currencyIDs: String) {
-        self.endpoint = URL(string: "https://651ff0cc906e276284c3c1bc.mockapi.io/api/v1/orders/1/payment/\(currencyIDs)")
-    }
-}
-
 final class PayDataProvider: PayDataProviderProtocol {
-
-    
-    
-    //static var shared = CardDataProvider()
-    
-    var networkClient: NetworkClient
-    
+    let networkClient: NetworkClient
     let orderChanged = Notification.Name("CartUpdated")
+    
+    struct CurrenciesRequest: NetworkRequest {
+        var endpoint: URL? = URL(string: "https://651ff0cc906e276284c3c1bc.mockapi.io/api/v1/currencies")
+    }
+    
+    struct PayRequest: NetworkRequest {
+        var endpoint: URL? = URL(string: "https://651ff0cc906e276284c3c1bc.mockapi.io/api/v1/orders/1")
+        init(currencyIDs: String) {
+            self.endpoint = URL(string: "https://651ff0cc906e276284c3c1bc.mockapi.io/api/v1/orders/1/payment/\(currencyIDs)")
+        }
+    }
     
     init(networkClient: NetworkClient = DefaultNetworkClient()) {
         self.networkClient = networkClient
     }
     
+    /// Функция возвращает массив валют [CurrencyDto]
     func getCurrencies(_ completion: @escaping (Result<[CurrencyDto], Error>) -> Void) {
         let currenciesRequest = CurrenciesRequest()
         networkClient.send(request: currenciesRequest , type: [CurrencyDto].self)  { result in
@@ -48,15 +39,14 @@ final class PayDataProvider: PayDataProviderProtocol {
                 switch result {
                 case let .success(data):
                     completion(.success(data))
-                    print("getCurrencies success. Count = \(data.count)")
                 case let .failure(error):
                     completion(.failure(error))
                 }
             }
         }
-        return
     }
     
+    /// Функция проводит оплату корзины с помощью валюты  currencyId
     func payOrder(currencyId: String, _ completion: @escaping (Result<PaymentDto, Error>) -> Void) {
         let paymentReqest = PayRequest(currencyIDs: currencyId)
         networkClient.send(request: paymentReqest , type: PaymentDto.self)  { result in
@@ -69,7 +59,6 @@ final class PayDataProvider: PayDataProviderProtocol {
                 }
             }
         }
-        return
     }
-    
+
 }
